backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
werkzeug==2.2.3
python-dotenv==1.0.0
sqlalchemy==2.0.23
python-pptx==0.6.21
email-validator==2.1.0
google-generativeai==0.3.1
requests>=2.28.0


